+++ MPChartLib/src/main/java/com/github/mikephil/charting/formatter/PercentFormatter.java
@@ -39,7 +39,7 @@ public class PercentFormatter implements IValueFormatter, IAxisValueFormatter
 
     // IAxisValueFormatter
     @Override
-    public String getFormattedValue(float value, AxisBase axis) {
+    public String getFormattedValue(double value, AxisBase axis) {
         return mFormat.format(value) + " %";
     }
