+++ MPChartLib/src/main/java/com/github/mikephil/charting/renderer/XAxisRendererHorizontalBarChart.java
@@ -144,7 +144,7 @@ public class XAxisRendererHorizontalBarChart extends XAxisRenderer {
             if (centeringEnabled) {
                 positions[i + 1] = mXAxis.mCenteredEntries[i / 2];
             } else {
-                positions[i + 1] = mXAxis.mEntries[i / 2];
+                positions[i + 1] = (float) mXAxis.mEntries[i / 2];
             }
         }
